bot updates
